refactor: HomeController 위치 변경 및 추가 작업
- 오타 수정
- DTO의 상속 삭제
