+++ src/main/java/roomescape/reservation/controller/HomeController.java
@@ -1,4 +1,4 @@
-package roomescape.home;
+package roomescape.reservation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
