+++ src/main/java/roomescape/reservation/dto/ReservationRequest.java
@@ -1,16 +1,22 @@
 package roomescape.reservation.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import roomescape.reservation.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class ReservationRequest extends Reservation {
-    public ReservationRequest() {
-        super();
-    }
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReservationRequest {
+    private String name;
+    private LocalDate date;
+    private LocalTime time;
 
-    public ReservationRequest(String name, LocalDate date, LocalTime time) {
-        super(null, name, date, time);
+    public Reservation toReservation() {
+        return new Reservation(null, name, date, time);
     }
 }
