+++ src/main/java/roomescape/reservation/dto/ReservationResponse.java
@@ -1,24 +1,30 @@
 package roomescape.reservation.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import roomescape.reservation.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class ReservationResponse extends Reservation {
-    public ReservationResponse() {
-        super();
-    }
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReservationResponse {
+    private Long id;
+    private String name;
+    private LocalDate date;
+    private LocalTime time;
 
     public ReservationResponse(Reservation reservation) {
-        super(reservation.getId(), reservation.getName(), reservation.getDate(), reservation.getTime());
+        this.id = reservation.getId();
+        this.name = reservation.getName();
+        this.date = reservation.getDate();
+        this.time = reservation.getTime();
     }
 
-    public ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
-        super(id, name, date, time);
-    }
-
-    public static ReservationResponse toEntity(Long id, ReservationRequest request) {
-        return new ReservationResponse(id, request.getName(), request.getDate(), request.getTime());
+    public static ReservationResponse fromReservation(Reservation reservation) {
+        return new ReservationResponse(reservation.getId(), reservation.getName(), reservation.getDate(), reservation.getTime());
     }
 }
